Updates
Hi Team,

I apologize to get back late. 

There are few changes I made to the doc which I feel should be merged. 

In addition, I want to mention the blue boxes that we have on the page are in wrong order. As I mentioned before, Step1 > Install, Configure and Verify and Step2 > Set-up Dev Environment while Step3 > Build your first application. I hope this can be corrected.

If there is something unclear in my above mail. Please reach out on 979-985-7127 or you can send me an email on hmadan@redhat.com or comment on this thread I would be notified.

I wish all of you a great weekend!

-Harshul